Internet gaming relay (#12)
Adds a new binary meant to run on the server called `wlnr` - widelands network relay. 

The `wlnr` listens on port 7397 for external connections - so this port must be opened in the firewall of widelands.org. The metaserver talks to `wlnr` via RPC and informs it of new games that should start. The relay accepts connections from players and relays messages between them - making it unnecessary for any player to mess with their firewall or router settings.

Closes #4.